improved error messages;
changed behaviour of builder:

the builder now tries to satisfy dependencies by running appropriate tools. If this is not possible, i.e., some file(s) cannot be produced, then a tool depending on that file may still be run. So each tool must make sure its input files exist and generate error messages if not. This was actually the purpose of the change of the behaviour: the tools can generate much better error messages.