Refactor Reasoner tests to check replay
Refactor test routines for succesful reasoner applications.  The new
code ensures that the reasoner can be replayed on some sequent created
with the hypotheses and goal present in the rule.  Moreover, it checks
that in such cases, the rule returned by the reasoner is very similar to
the original rule.  These new tests allow do detect some cases where a
reasoner generated a too liberal rule that would not pass replay (which
is considered a bug of the reasoner).

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=9349