Fix rules needing WD strictness.
Fix reasoners that should have checked for WD strictness and did not.  Also
added a check for existence of given hypothesis when it was missing.  This
introduces possible regression in case the reasoner would have been applied
directly, rather than through the user interface (very unlikely).

The reasoners modified are:
  org.eventb.core.seqprover.funInterImg
  org.eventb.core.seqprover.funOvr
  org.eventb.core.seqprover.funSetMinusImg

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=9787