Fix bug #3085103: Unsound behavior of newPP.
The bug was caused by an equivalence clause that was wrongly reported as a
contradiction when all its literals got canceled out.

I've fully checked that all cases of simplification of an equivalence
clause are now appropriately covered.

Added a regression test for the original bug + new tests for full
coverage for equivalence clause simplification.

The similar simplification for disjunctive clauses has also been checked
and found correct.

svn path=/trunk/RodinCore/org.eventb.pp.tests/; revision=10003