Fix bug 3102775.
Fix bug by strengthening condition for simplifying predicates of form
"t = TRUE".  The new condition asks that "t" is a constant to avoid
capture of variable during simplification.

This fix has a side-effect of making newPP maybe less efficient (although
it does not show in acceptance tests).

Commented out some unit tests that were relying on the previous version of
the simplification and that do not pass anymore.

svn path=/trunk/RodinCore/org.eventb.pp.tests/; revision=10210