Fix bug 3106076: Missing POs in interface.
Added checks for errors when fetching proof skeletons from proof files
and isolating such faults so that they do not propagate and break POM.
Done the same work for the proving UI (i.e., in UserSupport).

In the longer term, it might be interesting to recover a partial skeleton,
ignoring only the subtree rooted to an ill-formed node.  Also, currently,
in case of error, the problem is only traced in the log but not displayed
in the GUI (as the current protocol doesn't allow to report such problems).

svn path=/trunk/RodinCore/org.eventb.core.tests/; revision=10261