Increment version of AutoRewritesL1 to 1.
Bug 3158594 : Rodin 2.0.1 provers are unsound showed
that a glitch was introduced in the auto rewriter L1,
in the implementation of rule SIMP_IN_COMPSET_ONEPOINT.

This bug has independently been fixed when moving to a new
implementation in AutoRewritesL2.  However, this ``positive''
regression must be traced by a change in the reasoner version for L1.

svn path=/branches/RodinCore/AutoRewriterL2/; revision=10568