Protect against reasoner error
Ensure that the serialized proof tree is well-formed, even in case of
error when serializing the input of a reasoner.

Ensure that the serialized proof tree is well-read, even in case of error when
deserializing the input of a reasoner or repairing the input.

Add corresponding tests.

There is not yet proper error reporting to the user, because this would
need to break the Event-B core API.  This job is postponed to Rodin 4.0
where we could implement a general strategy.  A first idea would be that
proof serialization would produce a multi-status and ensure that the
serialized tree is deserializable (no dangling half-saved proof tree).
Errors should be reported as a marker so that the user is aware that
some problem occured in automated proof.