Extract predicate simplifier from the auto rewriter
Extract code that performs simple syntactic rewriting in predicates out of the
auto rewriter.  This will allow to apply the corresponding simplification rules
independently from the zillion other rules of the auto rewriter.

The extraction is done by creating a new super-class of AutoRewriterImpl named
PredicateSimplifier and putting in the new super-class the Tom implementation
of the rewriting rules to extract.

Also, to ease adaptation of the PredicateSimplifier to specific needs, an
option mechanism has been added to this class, which allows to specify more
finely which rewrite rules are applicable.  The level mechanism of the
autoRewriter translates into different sets of predicate simplifyer options.

The tests have been refactored similarly, extracting a new superclass named
PredicateSimplifierTests.

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=12402