Enable next and previous actions in prover UI
Enable global actions NEXT and PREVIOUS in the prover UI views and editor.
Also add missing updateActionBars() to signal the change to Eclipse.
Note: This change is not effective for the editor, as the editor page is not
properly registered.

See https://sourceforge.net/tracker/?func=detail&aid=3383270&group_id=108850&atid=651669 .

svn path=/trunk/RodinCore/org.eventb.ui/; revision=12649