Ensure regression tests for AutoRewriter are passed
The regression tests for previous levels of the AutoRewriter now run for
all levels, so that regressions can be detected and analysed if needed.
This is done by using subclassing. This also needed to change the name
of static fields of test classes, so that they don't mask the instance
field of the common superclass.
Also fix regression of L2 tests when running L3 and new rewritings that
were not properly protected with a level constant.

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=13143