Fix version of PP translator for Rodin 2.3.0
The version of the PP translator plug-in was incorrectly increased twice
since Rodin 2.2.0 has shipped.  The final version for Rodin 2.3.0 is 0.5.0,
not 0.6.0.
Also remove temporary method reduceToPredicateCalulus() which is not used
and fix tests accordingly.
Finally, also fix some tests that did not type-check (the problem was not
detected previously).

svn path=/trunk/RodinCore/org.eventb.pptrans.tests/; revision=13162