Refresh hypotheses views only when visible
Refresh hypotheses views only when they are visible.  Also, if
refresh was skipped, it is done when the view becomes visible.

As a consequence, one must not test on the existence of the styled text,
but on that of the view control to launch a refresh when the user
support changes (as the view might initially be invisible and the styled
text thus never created).

svn path=/trunk/RodinCore/org.eventb.ui/; revision=13440