Make TacticUIRegistry immutable
The class is now thread-safe because it is immutable. There is no need
anymore to make all query methods synchronized.

The analysis of the extension point has been moved to a separate class
to allow for easy refactoring and unit testing of it.

Also, extract remaining non-static subclasses to new files, at the
price of temporary code duplication (marked with FIXME).