Remove local commands from UI tactic registry.
The concept of command with a non-global target (i.e., a local command)
does not make sense as a command applies globally to a proof tree node
and not to a predicate.  I have therefore made explicit this constraint
in the API documentation and removed all code that was dealing with this
case.

Also, now verifies that tactics or commands contributed to a dropdown
or toolbar have a global target, as dropdowns and targets are in the
Proof Control view. If this is not the case, then the dropdown or
toolbar attribute is ignored.