Replace Strings with descriptive objects in proof UI
The proof UI was using Strings for dealing with toolbars, dropdowns,
global tactics and proof command when it could use real descriptor
objects.  This is fixed by this commit for the tactics and commands with
a global target.  Local tactics (i.e., applicable to a hypothesis or
goal) will be fixed similarly in another commit.