Fix cannot delete file error on Windows
When a Rodin file contained invalid XML, it was not properly closed.
On Unix (Linux and Mac OS X), this was benign (just keeping a dead file
descriptor).  On Windows, however, this translates in having the Java
VM keep a handle on the file, which prevents its deletion.

This is now fixed by always closing the file, whatever the result
returned by the XML parser.

svn path=/trunk/RodinCore/org.rodinp.core/; revision=14010