Improve wrapper used by XProver API
The previous version was systematically killing the external prover
with a SIGKILL signal, not leaving it any chance to clean up.  This
new improved version first sends a SIGTERM signal, then a SIGKILL
signal if the external prover did not stop in 100 ms.
Also add a test for command wrapping which has been added previously
to the API.

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=14018