Fix bug #3469348: Bug in INITIALISATION FIS PO
Ensure that the after values of the concrete variables never occur
in the WD and FIS proof obligations of a concrete event. See
http://wiki.event-b.org/index.php/Revisiting_Feasibility_POs for a
detailed analysis of the bug.

The code change consists in putting the abstract actions in
hypothesis only if no parameter witness contains a primed variables.
Moreover, the variable witnesses are never put in hypothesis to
prevent any circular argument.

A regression test for the bug has been added and all other tests for
FIS proof obligations have been carefully checked and possibly fixed
to fit the new proof obligation generation algorithm.

svn path=/trunk/RodinCore/org.eventb.core.tests/; revision=14090