Cleanup the extension point description of proofTactics
Changed part of the descriptions, fixed typos, added example of "any"
tactic contribution.

The use of class DefaultTacticProvider was mistakenly indicated as
required in the extension point schema. However, this default
implementation is just a help and it is not mandatory to subclass it
(actually the org.eventb.ui plug-in itself ignores it!)

svn path=/trunk/RodinCore/org.eventb.ui/; revision=14273