Push further tactic application proxys
Push further the proxys created to encapsulate tactic applications.
Proxys are now available almost everywhere in the UI code and advantage
of this can be taken in a coming refactoring step.

Also merge methods getTacticApplicationsToGoal() and
getTacticApplicationToHypothesis() which where very similar and
always called within an if statement testing for the goal or hypothesis
case.

Remove now useless method retainPredicateApplication().