Work around problem raised by API tools for org.eventb.ui-2.5.0
With the publication of Rodin 2.5 RC1, the API problem filters have been
removed.  However, Eclipse 3.7.2 finds problems with classes supposedly
changed in fragment org.eclipse.ui.cocoa. However, this classes are exactly
the same in Eclipse 3.7.1 and 3.7.2.  I think that the problem is rather
caused by the presence of a fragment that the API tooling does not handle
properly.

This commit works around this by adding new problem filters to the
org.eventb.ui plug-in.

svn path=/trunk/RodinCore/org.eventb.ui/; revision=14575