Fix bug #3454414: Problems when merging events
When verifying the witnesses of a concrete event that merges several abstract
events, only the parameters of the first abstract event were entered in the
type environment.  This caused strange error messages when checking the witness
for an abstract parameter that occurs only in the second or later abstract event.

This commit fixes this bug, taking care of the case where several abstract
parameters have the same name but different types (they therefore could not
be entered in the same type environment).

In the end, this fix is not complete, but good enough.  A complete fix would
be to check witnesses separately for each refinement link of the concrete
event, and then check that all witnesses have been covered (i.e., every
witness has been checked with respect to at least one abstract event).

svn path=/trunk/RodinCore/org.eventb.core.tests/; revision=14604