Fix bug #3521058: dynamic label for open preferences with scope not correct
We need to listen to the changes to a specific preference in the project
scope to detect when project specific preferences are enabled or
disabled.

To work more cleanly, I've started by creating a new separate class
OpenPreferencesAction for the action that opens the preference dialog.  This
allows to completely separate this action from the ProofControlPage and
removes some clutter in ProofControlPage.

The new class OpenPreferencesAction implements a listener on the project
scope tactic preferences to detect when their enableness changes.

Move method hasProjectSpecificTactics() from class TacticPreferenceUtils to
OpenPreferencesAction as it was used only in the latter class.

svn path=/trunk/RodinCore/org.eventb.ui/; revision=14625