Add new method for instiantiating a parameterized tactic
Clients were allowed to specify only the id of the new instance, but not
its name, nor its description which were inherited from the parameterized
tactic.  Consequently, the user was faced with several instances all with
the same name, which is difficult to understand.

This commit adds a new version of the instantiation method that allows
clients to specify the name and description of the new instance.

Also add corresponding tests to ensure that the additional information
is properly propagated.

This is a backward compatible API change.

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=14628