Add factory method to create an empty set extension
The AST library did not provide any mean for creating a set extension
node with no children but with a given type.  One had to create an
untyped set extension, and then to type-check it in some environment
to define its type.

This commit fixes this by providing a new factory method for that.

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=14942