New implementation of typed rewriting
The point of this commit is to prepare for more general rewriters that
can change at the same time the mathematical language, the type and the
nodes of some formula.

Moved the checkReplacement() methods to static methods (they did not
depend on any object state) and use them everywhere needed.

Moved all type-checks during rewriting to the rewriter wrappers (rather
than in the sub-classes of Formula).

Change the interface for rewriting mathematical extensions, preparing
for the case where the mathematical language would also change during
rewriting.

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=14944