Rename and move interface ITypedFormulaRewriter
This interface is only for internal use and must not be part of the
API. It has therefore been moved to an internal package. Moreover,
I've renamed it to ITypeCheckingRewriter which conveys better the
responsibility for implementers of this interface.

Also, rename some implementations of this interface to convey better the
role of each of them.

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=14945