Fix bug #3488583: Duplicate PO for action WD
The bug fix in Subversion commit 14090 was two strong for action WD
proof obligations.  Consequently, many action WD proof obligations
were generated although the abstract event contained the exact same
action which had already been proved well-defined.

As the only identifiers that can occur in an action are either constants,
variables or parameters and they cannot change meaning between an
abstract and a concrete event, we can inherit the WD proof from the
abstraction (which is not the case in general for feasibility).

Therefore, it is valid to make a special case for action WD generation,
and I have reverted the condition for generating such a PO to the 
previous test: generate this proof obligation only if the abstract
event does not contain the exact same action.

I've also added a regression test specific to this bug which corresponds
to the project which was attached to the bug report.

svn path=/trunk/RodinCore/org.eventb.core.tests/; revision=14963