Introduce target language in math extension translation
When creating fresh names during translation of a mathematical extension,
it is the factory that contains the original datatype which is used. It
is rather a factory without the datatype that should be used.  This
would correspond to a formula factory for the target mathematical
language.

As the full machinery for math extension translation is not implemented
yet, I've just made a work-around by using a hard-coded target language,
which allows the tests to pass.  Put a FIXME to remember to fix this in
the near future.

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=14967