Separate formula factory and type environment in tactic tests
As most tests use the default formula factory, it is useless to set it
in every test.  Moreover, this new setting avoids creating two instances
of a type environment and DLib for each test.

Made all methods of AbstractTacticTests protected to convey that they
are meaningful only to sub-classes implementing tests.

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=14980