Fix bug in tactic InDomGoal
The code was looking for all "dom" sub-expressions, but the tactic can
only deal with the one occurring at the right-hand side of a membership
predicate.  This prevented the tactic from succeeding when a dom
expression occurred in the left-hand side of the membership.  This has
been fixed by hard-coding the position of the dom sub-expression.

Add new regression test with dom on the left-hand side.

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=14991