Hide warnings in sequent prover tests
The tests for the external prover framework are using the deprecated
interface XProverCall.  However the implementation of this interface is
common with the new interface XProverCall2.

As both interfaces will get merged in Rodin 3.0, we just hide the
deprecation warnings, knowing that the tests will be rewritten at that
time.

svn path=/trunk/RodinCore/org.eventb.core.seqprover.tests/; revision=15004