Fix leak in TypeEnvironment implementation
Method Map#keySet() returns a view on the internal data structure of
a map.  Consequently, a client could change a type environment as
a side effect of modifying the set returned by
ITypeEnvironment#getNames().

This has been fixed by returning the set encapsulated into an
unmodifiable set to prevent such behavior.  The set nevertheless
stays synchronized with the type environment representation, so
that modifications to the type environment are reflected into the
returned set, as before.

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=15103