Remove class ArgProduct in AST library
The AST library does not have the notion of a relational type.  There
is a makeRelationalType(left, right) factory method, but internally
this is translated to makePowerset(makeProduct(left, right)).

I've therefore removed this class which should not have been created
to start with.  It could even be harmful as it will not equal the
equivalent definition based on PowerSet and Product.

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=15106