Implement new method IDatatype.getArgumentSets()
This method returns the list of argument sets for some datatype
constructor, given a datatype set, that is a set created with the
datatype type constructor.

Add corresponding acceptance tests. As a side-effect a few bugs where
fixed in class InjectedDatatypeExtension (choice of factory, parsing of
non parametric datatype).

This is a joint work of Thomas Muller and Laurent Voisin.

svn path=/trunk/RodinCore/org.eventb.core.ast.tests/; revision=15119