Ensure validity of identifier names in AST library
Ensure that all AST nodes that contain a name are checked for valid name
upon creation. This condition is necessary to fulfill the library
invariant that the serialization of a Formula (with a toStringXXX()
method) can be parsed back into the exact same formula.

The affected nodes are:
  - BoundIdentDecl
  - FreeIdentifier
  - GivenType
  - Predicate Variable

I have checked all occurrences of String in class FormulaFactory to make
sure that there is no other unchecked use of strings in the AST node
factory.

Add corresponding tests in class TestFormulaFactory.

API note: The constructors of some AST nodes have changed.  The API
tooling of Eclipse considers that it is an API breakage.  I disagree
because these constructors are not part of the API: They are protected
and their class cannot be extended by clients.  Therefore, I've added
the corresponding API filters.

svn path=/trunk/RodinCore/org.eventb.core.ast.tests/; revision=15123