Rewrite bound declarations in the AST library.
Initially, bound identifier declarations of quantified formulas were
never rewritten, because the only useful information in a declaration
is its type which never changed.

However, with the introduction of type-checking rewriter which can also
rewrite types, it becomes necessary to also rewrite bound identifiers.
In the implementation of specialization and datatype translation, this
rewriting of declarations was done inappropriately by implementing
specialized rewrite() methods on quantified formulas.

This commit fixes this by implementing the rewriting of declarations
directly into the classes of quantified formulas.

No test has been added, because we can rely on existing tests (e.g.,
TestDatatypeRewriter).

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=15186