Fix access to private field in AST library
The Oracle Java compilers (both 1.6 and 1.7) choke on an access to the
private field "hashCode" in some static method of class Formula.
Although this access is valid according to the Java Language
Specification, it is the only case where we're not using the public
accessor method "hashCode()" so I've changed the code to be more
consistent.

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=15226