Code cleanup in type environments
Add comments for design decisions.

Remove all "non-Javadoc" comments (these comments have a strong
tendency to become out of date, are never fixed and useless as Eclipse
provides hyperlinks to the super-class or interface).

Remove duplicate implementation of hashCode().  It is OK if two type
environments have the same hash code and are not equal because one is
mutable and not the other.

svn path=/trunk/RodinCore/org.eventb.core.ast/; revision=15355