Remove useless formula factory parameter in Event-B core API.
Some methods of the database customization layer for Event-B were
asking for both a formula factory and type environment parameter.
However, the type environment already has a formula factory, hence
it is useless and cumbersome to pass it again explicitly.

Moreover, there was a risk of discrepancy between the formula factory
parameter and the type environment, if the latter was not built by
the same formula factory.  This was not checked by the current code.

By removing the formula factory parameter, we remove the risk and make
the API simpler.

svn path=/trunk/RodinCore/org.eventb.core.tests/; revision=15361