Fix invalid type in DatatypeTranslator.
When constructing a set built with the datatype type constructor, we
were passing a wrong type: it was the datatype instance instead of a
power set of it.  However, as we are building with the datatype type
constructor, we know that the type can always be synthesized from the
children, so it is better to not pass any type to
makeExtendedExpression() and let type synthesis work normally.

This bug was not detected earlier, because the resulting node is never
used directly, but only for harvesting the set arguments from the
constructed set.