Ensure that the type given at node creation is valid
Add an assertion at the end of a node construction to check that
provided type has been correctly set (it could not have been done if the
given type was incompatible).

Remove test illTypedExtendedExpression which is no longer necessary
because makeExtendedExpression now throws an exception.

Creating a free identifier with an illegal type is now expected to fail.