Fix bug #635: Inconsistent result of formula type-checking
By design, the type-checking result should be consistent with the
type-checked state of the root formula.  The only case where they can be
different is when the type environment passed to the type-checker is
incompatible with the formula.

To ensure that this property is indeed implemented by the AST code, we
add an assertion stating it before returning the type-check result.
Then, if another occurrence of this bug happens, it will be detected.

Add regression test where type-check fails, although formula is
type-checked (caused by an incompatible type environment).