Fix typecheck issues with inconsistent name of given types
The type checking problems due to identifiers used at the same time
for given sets and other types are now detected and reported during the
typecheck() procedure by verifying types of nodes that could introduce
new given sets (it detects a problem that was occurring during last
step of Formula.solvingTypes() whereas it should not).

As a consequence in core code:

- TypeCheckResult: the given sets declared in given types are now added
to the resulting inferred type environment.

- Type: the use of Type.getGivenTypes() procedure does not require the
type to be solved anymore and returns a best effort result in case the
type is not solved (a TypeVariable cannot add any given type since it is
not a solved type): this modification was mandatory in case
of ExtendedExpression because an unsolved type could introduce a
new given set.

Modify tests in consequence:

- TestTypeChecker: in a test the inferred type environment now contains
a given set declared in given types

- IdentsChecker: only names of free identifiers are checked when the
type check failed before. It allows to do the verification when a free
identifier name has several types in the formula and as consequence
the type checking fails (the free identifiers cache keeps only one of
the types).