Fix typecheck issues with inconsistent name of given types
The type checking problems due to identifiers used at the same time
for given sets and other types are now detected and reported during the
typecheck() procedure by verifying types of nodes that could introduce
new given sets (it detects a problem that was occurring during last
step of Formula.solvingTypes() whereas it should not).

As a consequence in core code:

- TypeCheckResult: the given sets declared in given types are now added
  to the resulting inferred type environment.

- Type: the use of Type.getGivenTypes() procedure does not require the
  type to be solved anymore and returns a best effort result in case the
  type is not solved (a TypeVariable cannot add any given type since it
  is not a solved type): this modification was mandatory in case of
  ExtendedExpression because an unsolved type could introduce a new
  given set.

Modify tests in consequence:

- TestTypeChecker:
  + the inferred type environment now contains a new given types
  + ensure that incompatible use of an identifier as a type and not a
    type are now detected.

- TestExprTypeChecker: Verify that the expected type is compatible
  with the given type environment. Ensure that incompatible typing of
  an identifier is detected. Update existing tests for the larger
  inferred type environment (now contains all given sets).

- IdentsChecker: only names of free identifiers are checked when the
  type check failed before. It allows to do the verification when a free
  identifier name has several types in the formula and as consequence
  the type checking fails (the free identifier cache keeps only one of
  the types).

When checking that a name denotes a given type, if the name already has
a type variable as type, we must unify this type variable with the
powerset of the given type to remember this restriction.  Add
corresponding regression test.