Refactor addition of given types to identifier cache
Externalize the extraction and sorting of free identifiers hidden in a
type to a new utility class. Replace the hand-coded sorting by a call to
existing sorting method.

Do not try to reconstruct a posteriori a formula factory, but reuse the
one that was passed to the constructor or type-checker.

Factor out common code in a new method of Expression.

No API change.