Fix synthesizeType() for quantified formulas
The original solution was incorrect: an identifier conflict detected
with all but the last bound identifier declaration was not reported.

Use an intermediate array to fetch free identifiers from all children
rather than repeated calls to makeMerger().  This corresponds better
with what is done in other similar classes and fixes this bug.

Add corresponding unit tests which also are regression tests for the
preceding incorrect behavior.