Refactor Formula.equals()
Now that equality without alpha conversion is not there anymore, we can
fully refactor the way equality of formula is computed. Beyond
homogenization of code, the most notable new features are:

- Compare the tags of the formulas in Formula.equals() rather than
in each subclass method. Moreover, this is the behavior which is
documented in the Javadoc of Formula.equals(Formula, boolean).

- Type compatibility is now checked in class Expression rather than in
every subclass of it.

- Use Arrays.equals() for recursive calls, rather than hand-coded
equivalent (the initial code was Java 1.4 which does not have class
java.util.Arrays).

- Remove other code duplications when possible (e.g., type comparison).

No functional change.