Forbid to build V1 node with a V2 FF (and resp.)
Add assertion to forbid the construction with a V1 tag (KPRJ1, KPRJ2 or
KID) in UnaryExpression if the FormulaFactory is not of V1 type.

Add assertion to forbid the construction with a V2 tag (KPRJ1_GEN,
KPRJ2_GEN or KID_GEN) in AtomicExpression if the FormulaFactory is not
of V2 type.

Fix FormulaFactory equals method by checking the grammar class is the
same since it was allowing to be equal with different grammars. Enforce
that a grammar is given in constructor.

Fix AST tests regarding node creations since it was building V1 nodes
with a V2 FormulaFactory.
Fix AST tests regarding node creations since it was building nodes with
a V2 FormulaFactory whereas those nodes were defined to be part of the
V1 language.