Build formulas with the right factory in AST tests
When building programmatically formulas in tests, a mix of V1 and V2
formula factories were used.  This made the trees heterogeneous and will
not work anymore soon.

After this commit, it is expected that all tests use the right factory
for building nodes, especially for operators that are specific to one
language version, such as non-generic projections and identity, and the
partition operator.