Make IdentsChecker raise errors
The class IdentsChecker used to return a boolean value stating whether
the identifier caches of a formula were correct.  However, this is very
cumbersome to debug when the check fails.

Now this class raises a failed assertion as soon as a problem is
discovered.

I have also cleaned up the code consequently.