Fix V1 to V2 Upgrader
VersionUpgraderV1V2 was implementing IFormulaRewriter to do the upgrade
whereas an formula rewriter must not modify the language. Now it
implements ITypeCheckingRewriter which is the correct level to modify
the language version.

ITypeCheckingRewriter interface has been modified since unary expression
intermediate node cannot be built during a V1 to V2 upgrade, as a
consequence we pass part of the internal data-structures that hold the
already rewritten children.

Delete useless UpgraderFactory since only one upgrader is possible.