FIXUP typecheck issues with inconsistent name of given types
Add a regression test that fails an assertion in the formula
type-checker: success although the formula is not type-checked.

Then, fix the issue: when checking that a name denotes
a given type, if the name already has a type variable as type, we must
unify this type variable with the powerset of the given type to remember
this restriction.