Fix default type checking rewriter
Since the factory used to build a node is accessible, we must build
again a leaf node if the factory used to rewrite is different (it will
propagate rebuild of parent nodes).

ITypeCheckRewriter java documentation now indicates that it is
mandatory to rebuild a leaf node if the rewriter factory used is
different than the node factory.

DefaultTypeCheckingRewriter class and subclasses are now rebuilding
leaf nodes if the factory is different from the original one.

BoundIdentDeclRemover does not implements ITypeCheckingRewriter
directly anymore since its indirect superclass
DefaultTypeCheckingRewriter did.