Cleanup factory for extended formulas
Protection against clients (e.g., cloning of arrays passed as argument)
is now done by the factory method rather than the constructor.  This
avoids creating twice the same array when clients use collections and
allows to now have only one constructor per class for all sub-classes
of Formula.

Speed up the check that an extension is supported (it was using linear
probing on the range of a map).

Homogenize the interface by adding a factory method for creating an
extended expression from collections and a type (it was missing).