Cleanup factory for extended formulas
Homogenize the interface by adding a factory method for creating an
extended expression from collections and a type (it was missing).

Protection against clients (e.g., cloning of arrays passed as argument)
is now done by the factory method rather than the constructor.  This
avoids creating twice the same array when clients use collections.

Refactor the verification of the extension to make it faster (it was
using linear probing on the range of a map).