Improve the API of FormulaFactory
Revisit the Javadoc of all methods of FormulaFactory that create new
Type and Formula instances.  Make more precise the pre-conditions on
parameters.

Note that the form of quantified expressions is now a pure decoration
and is always upgraded automatically if needed.

This defines the new API contract with clients, which is not yet
enforced.