Implement API for extended formula creation
Implement and test the API for extended formula creation.

During implementation, I've discovered that a strange expression
operator could wreak havoc in the type-checker algorithm. The issue is
that the type-checker infers a correct type for an extended expression,
but the expression itself does not type-check for some other reason.
This behavior never occurs with regular operators and is a special
novelty brought by extensions.  I've added a regression test for such
a situation.

It is on purpose that the assertion about the nature of the extension
(expression or predicate) has been left as an assertion and not changed
to throw an exception.  In fact, this check shall be done by static
checking and not here.  This will be fixed when we revisit the way
extensions are contributed to formula factories.