Fix datatype creation in tests
Class InjectedDatatypeExtension was parsing wrongly the types of
constructor arguments, creating given types instead of references to
type parameters.  This is fixed by this commit.

However, I think that this argument type parsing should be part of the
AST library, rather than buried into tests.  This will be revisited
later with the whole API for datatype creation.

This fix also removes the dependency on a formula factory, as the
factory to use for parsing argument types is supplied by the mediator
for datatype construction.