Fix datatype creation in tests
Fix invalid datatype specification in testBug3574565().  Method
makeDatatypeFactory() did not know how to create two datatypes from one
String specification, and the resulting factory was all wrong.

Simplify code in InjectedDatatypeExtension:
  - There is no need to have a special case for product types.
  - Add a comment explaining the kludge for inductive types.
  - Simplify the regular expression for inductive types.
  - Remove FormulaFactory argument of the constructor.

The last change is justified by the fact that the factory to use for
parsing argument types is supplied by the mediator for datatype
construction.  It was indeed a bad idea to use any other factory.