Remove assertion in Expression.setFinalType()
It could happen that the proposed type had been computed by the
type-checker but would be wrong because of some type-check problem:
we do not fully master the chains of unifications of type variables and
a far-reached unification problem is not always propagated back to all
involved type variables because of path compression.

We therefore remove the assertion about the equality of synthesized and
proposed type and just silently ignore the issue here.

This change is not ideal, as it could mask an issue in the type-checker,
but I do not know how to do it differently.