Verify that type-check status is hierarchical
A strong invariant of the AST library is that, once a node is marked as
type-checked, it is fully immutable.  To ensure this, we need that when
a node is marked as type-checked, so are all of its descendants,
otherwise their type could change thus mutating the tree.

Add a new check in the AST tests to verify this invariant.

Also add a regression test for the case of an expression extension that
does not verify that all its children are type-checked before accepting
that the node is type-checked.

This test has shown a serious flaw in the AST code where the expression
extension was trusted to perform the check on children in method
verifyType().  Now, this check is performed by the AST library itself.