Fix bug in QuantifiedExpression.getChild()
This method was not checking completely the given index and could raise
an ArrayIndexOutOfBoundsException when given a negative index. This is
fixed by this commit.

Instrument regression tests for sub-formulas to verify that there is no
other similar case.

Also, a few related improvements that do not introduce any functional
change:
  - Remove redundant check in ExtensionHelper.
  - Make methods final in Identifier.